move deconcat
