+++ nilearn/glm/second_level/second_level.py
@@ -31,12 +31,14 @@ from nilearn.glm.first_level.design_matrix import (
 from nilearn.glm.regression import RegressionResults, SimpleRegressionResults
 from nilearn.image import mean_img
 from nilearn.maskers import NiftiMasker, SurfaceMasker
-from nilearn.maskers._utils import (
-    deconcatenate_surface_images,
-)
 from nilearn.mass_univariate import permuted_ols
-from nilearn.surface import SurfaceImage, check_same_n_vertices, concat_imgs
-from nilearn.surface import mean_img as surf_mean_img
+from nilearn.surface.surface import (
+    SurfaceImage,
+    check_same_n_vertices,
+    concat_imgs,
+    two_dim_to_one_dim,
+)
+from nilearn.surface.surface import mean_img as surf_mean_img
 
 
 def _input_type_error_message(second_level_input):
@@ -331,7 +333,7 @@ def _get_con_val(second_level_contrast, design_matrix):
 def _infer_effect_maps(second_level_input, contrast_def):
     """Deal with the different possibilities of second_level_input."""
     if isinstance(second_level_input, SurfaceImage):
-        return deconcatenate_surface_images(second_level_input)
+        return two_dim_to_one_dim(second_level_input)
     if isinstance(second_level_input, list) and isinstance(
         second_level_input[0], SurfaceImage
     ):
