+++ nilearn/maskers/_utils.py
@@ -1,9 +1,6 @@
-import copy
-
 import numpy as np
 
 from nilearn import image
-from nilearn.surface import SurfaceImage
 
 
 def _check_dims(imgs):
@@ -41,43 +38,3 @@ def concat_extract_surface_data_parts(img):
         Concatenated data across hemispheres.
     """
     return np.concatenate(list(img.data.parts.values()), axis=0)
-
-
-def deconcatenate_surface_images(img):
-    """Deconcatenate a 3D Surface image into a a list of SurfaceImages.
-
-    Parameters
-    ----------
-    img : SurfaceImage object
-
-    Returns
-    -------
-    :obj:`list` or :obj:`tuple` of SurfaceImage object
-    """
-    if not isinstance(img, SurfaceImage):
-        raise TypeError("Input must a be SurfaceImage.")
-
-    if len(img.shape) < 2 or img.shape[1] < 2:
-        return [img]
-
-    mesh = img.mesh
-
-    return [
-        SurfaceImage(
-            mesh=copy.deepcopy(mesh),
-            data=_extract_surface_image_data(img, i),
-        )
-        for i in range(img.shape[1])
-    ]
-
-
-def _extract_surface_image_data(surface_image, index):
-    mesh = surface_image.mesh
-    data = surface_image.data
-
-    return {
-        hemi: data.parts[hemi][..., index]
-        .copy()
-        .reshape(mesh.parts[hemi].n_vertices, 1)
-        for hemi in data.parts
-    }
